requirements.txt
opencv-python-headless
pillow
ultralytics==8.0.196
insightface==0.7.3
onnxruntime==1.17.1
cloudinary
flask
gunicorn
uvicorn
python-dotenv
pymongo[srv]
numpy==1.26.4
scipy
dnspython